Prep SideNav redesign
