+++ packages/components/scripts/dev-server/site.tsx
@@ -3,28 +3,15 @@ import { render } from "react-dom"
 import { ThemeProvider } from "glamorous"
 import { injectStylesheet, baseStylesheet } from "@operational/utils"
 
-import { ContextMenu, ContextMenuItem, Icon, contiamoTheme } from "../../src/index"
+import { Sidenav, SidenavHeader, SidenavItem, contiamoTheme } from "../../src/index"
 
 class Site extends React.Component<{}, {}> {
   render() {
     return (
       <ThemeProvider theme={contiamoTheme}>
-        <div style={{ padding: 40 }}>
-          <ContextMenu>
-            <span>
-              <Icon name="X" size={30} />
-            </span>
-            <ContextMenuItem
-              onClick={() => {
-                console.log("clicked")
-              }}
-            >
-              Menu 1
-            </ContextMenuItem>
-            <ContextMenuItem>Menu 2</ContextMenuItem>
-            <ContextMenuItem>Menu 3</ContextMenuItem>
-          </ContextMenu>
-        </div>
+       <Sidenav expandOnHover> 
+        <SidenavHeader label="Hello"/>
+       </Sidenav>
       </ThemeProvider>
     )
   }
