--- packages/components/src/SideNavigation/Item/__tests__/SideNavigationItem.test.tsx
@@ -1,12 +1,0 @@
-import * as React from "react"
-import { render } from "enzyme"
-
-import ThemelessSideNavigationItem from "../SideNavigationItem"
-import wrapDefaultTheme from "../../../utils/wrap-default-theme"
-
-const SideNavigationItem = wrapDefaultTheme(ThemelessSideNavigationItem)
-
-test("SideNavigationItem component renders", () => {
-  const output = render(<SideNavigationItem>Hi, I'm an Item</SideNavigationItem>)
-  expect(output).toMatchSnapshot()
-})
