--- packages/components/src/SideNavigation/__tests__/SideNavigation.test.tsx
@@ -1,11 +1,0 @@
-import * as React from "react"
-import { shallow } from "enzyme"
-
-import wrapDefaultTheme from "../../utils/wrap-default-theme"
-import ThemelessSideNavigation from "../SideNavigation"
-
-const SideNavigation = wrapDefaultTheme(ThemelessSideNavigation)
-
-test("SideNavigation component renders", () => {
-  expect(shallow(<SideNavigation />)).toMatchSnapshot()
-})
